+++ src/logging.h
@@ -528,6 +528,7 @@ to_string (detail::parameter_t<detail::query_ref<T>, name, kind> param)
   F (amd_dbgapi_log_level_t)                                                  \
   F (amd_dbgapi_memory_precision_t)                                           \
   F (amd_dbgapi_alu_exceptions_precision_t)                                   \
+  F (amd_dbgapi_group_segment_out_of_addr_range_exception_t)                  \
   F (amd_dbgapi_os_queue_type_t)                                              \
   F (amd_dbgapi_process_id_t)                                                 \
   F (amd_dbgapi_process_info_t)                                               \