+++ src/os_driver.cpp
@@ -65,6 +65,8 @@ one_os_process_flag_to_string (os_process_flags_t flag)
       return "PRECISE_MEMORY";
     case os_process_flags_t::precise_alu_exceptions:
       return "PRECISE_ALU_EXCEPTIONS";
+    case os_process_flags_t::group_segment_out_of_addr_range_exception:
+      return "GROUP_SEGMENT_EXCEPTIONS";
     }
 
   return to_string (
@@ -162,8 +164,9 @@ to_string (os_agent_info_t os_agent_info)
     ".debugging_supported=%d, .address_watch_supported=%d, "
     ".address_watch_register_count=%zd, .address_watch_mask_bits=%#" PRIx64
     ", .watchpoint_exclusive=%d, .precise_memory_supported=%d, "
-    ".precise_alu_exceptions_supported=%d, .firmware_supported=%d, "
-    "ttmps_always_initialized=%d }",
+    ".precise_alu_exceptions_supported=%d, "
+    ".group_segment_out_of_addr_range_exception_supported=%d,"
+    " .firmware_supported=%d, ttmps_always_initialized=%d }",
     os_agent_info.os_agent_id, os_agent_info.name.c_str (),
     os_agent_info.domain, os_agent_info.location_id, os_agent_info.gfxip[0],
     os_agent_info.gfxip[1], os_agent_info.gfxip[2], os_agent_info.simd_count,
@@ -182,6 +185,7 @@ to_string (os_agent_info_t os_agent_info)
     os_agent_info.address_watch_mask_bits, os_agent_info.watchpoint_exclusive,
     os_agent_info.precise_memory_supported,
     os_agent_info.precise_alu_exceptions_supported,
+    os_agent_info.group_segment_out_of_addr_range_exception_supported,
     os_agent_info.firmware_supported, os_agent_info.ttmps_always_initialized);
 }