+++ src/os_driver_kfd.cpp
@@ -675,6 +675,9 @@ kfd_driver_base_t::agent_snapshot (
       agent_info.precise_alu_exceptions_supported
         = (entry.capability
            & HSA_CAP_TRAP_DEBUG_PRECISE_ALU_OPERATIONS_SUPPORTED);
+      agent_info.group_segment_out_of_addr_range_exception_supported
+        = (entry.capability2
+           & HSA_CAP2_TRAP_DEBUG_LDS_OUT_OF_ADDR_RANGE_SUPPORTED);
       agent_info.firmware_supported
         = entry.capability & HSA_CAP_TRAP_DEBUG_FIRMWARE_SUPPORTED;
       agent_info.address_watch_mask_bits = utils::bit_mask (
@@ -1917,6 +1920,9 @@ kfd_driver_t::set_process_flags (os_process_flags_t flags) const
     args.set_flags.flags |= KFD_DBG_TRAP_FLAG_SINGLE_MEM_OP;
   if (!!(flags & os_process_flags_t::precise_alu_exceptions))
     args.set_flags.flags |= KFD_DBG_TRAP_FLAG_SINGLE_ALU_OP;
+  if (!!(flags & os_process_flags_t::group_segment_out_of_addr_range_exception))
+    args.set_flags.flags |= KFD_DBG_TRAP_FLAG_LDS_OUT_OF_ADDR_RANGE;
+
 
   int err = kfd_dbg_trap_ioctl (KFD_IOC_DBG_TRAP_SET_FLAGS, &args);
   if (err == -ESRCH)